% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialrjars.R
\docType{package}
\name{dialrjars-package}
\alias{dialrjars}
\alias{dialrjars-package}
\title{Package required libphonenumber jars for the dialr package}
\description{
The dialrjars package collects libphonenumber jars required for the dialr package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/socialresearchcentre/dialrjars}
  \item Report bugs at \url{https://github.com/socialresearchcentre/dialrjars/issues}
}

}
\author{
\strong{Maintainer}: Danny Smith \email{danny@gorcha.org}

Other contributors:
\itemize{
  \item Google Inc. (libphonenumber) [contributor, copyright holder]
  \item Vonage Holdings Corp. (libphonenumber) [contributor, copyright holder]
  \item Ian Galpin (libphonenumber) [contributor, copyright holder]
  \item Ben Gertzfield (libphonenumber) [contributor, copyright holder]
}

}
\keyword{internal}
